+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Fixes
 
+- crashedLastRun now returns the correct value ([#4829](https://github.com/getsentry/sentry-react-native/pull/4829))
 - Fixes Feedback Widget accessibility issue on iOS ([#4739](https://github.com/getsentry/sentry-react-native/pull/4739))
 
 ## 6.14.0